# client/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Resume Builder API
  version: 0.1.0
  description: AI-powered resume builder — resumes, auth, and Claude-assisted content.
paths:
  /api/ai/generate-summary:
    post:
      operationId: generate_summary
      tags:
      - ai
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateSummaryRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GenerateSummaryRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GenerateSummaryRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateSummaryResponse'
          description: ''
  /api/ai/improve-bullet:
    post:
      operationId: improve_bullet
      tags:
      - ai
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImproveBulletRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ImproveBulletRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImproveBulletRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImproveBulletResponse'
          description: ''
  /api/ai/parse-resume:
    post:
      operationId: parse_resume
      tags:
      - ai
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseResumeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ParseResumeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ParseResumeRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResumeResponse'
          description: ''
  /api/ai/tailor-jd:
    post:
      operationId: tailor_jd
      tags:
      - ai
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TailorJDRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TailorJDRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TailorJDRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TailorJDResponse'
          description: ''
...
components:
  schemas:
    AIUsage:
      type: object
      description: Per-user AI usage snapshot (limits enforced in Phase 7).
      properties:
        calls_today:
          type: integer
        calls_this_month:
          type: integer
        daily_limit:
          type: integer
        monthly_limit:
          type: integer
        is_admin:
          type: boolean
      required:
      - calls_this_month
      - calls_today
      - daily_limit
      - is_admin
      - monthly_limit
...
    Education:
      type: object
      properties:
        id:
          type: string
          maxLength: 64
        school:
          type: string
          maxLength: 120
        degree:
          type: string
          maxLength: 120
        field:
          type: string
          maxLength: 120
        graduationDate:
          type: string
        gpa:
          type: string
          maxLength: 10
      required:
      - id
...
    GenerateSummaryRequest:
      type: object
      properties:
        content:
          $ref: '#/components/schemas/ResumeContent'
      required:
      - content
    GenerateSummaryResponse:
      type: object
      properties:
        summary:
          type: string
        ai_usage:
          $ref: '#/components/schemas/AIUsage'
      required:
      - ai_usage
      - summary
    ImproveBulletRequest:
      type: object
      properties:
        text:
          type: string
          maxLength: 2000
        role:
          type: string
          maxLength: 120
      required:
      - text
    ImproveBulletResponse:
      type: object
      properties:
        suggestion:
          type: string
        ai_usage:
          $ref: '#/components/schemas/AIUsage'
      required:
      - ai_usage
      - suggestion
    ParseResumeRequest:
      type: object
      properties:
        text:
          type: string
          maxLength: 20000
      required:
      - text
    ParseResumeResponse:
      type: object
      properties:
        content:
          $ref: '#/components/schemas/ResumeContent'
        ai_usage:
          $ref: '#/components/schemas/AIUsage'
      required:
      - ai_usage
      - content
...
    PersonalInfo:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        headline:
          type: string
          maxLength: 120
        email:
          type: string
          format: email
        phone:
          type: string
          maxLength: 40
        location:
          type: string
          maxLength: 100
        linkedin:
          type: string
          format: uri
        github:
          type: string
          format: uri
        website:
          type: string
          format: uri
    Project:
      type: object
      properties:
        id:
          type: string
          maxLength: 64
        name:
          type: string
          maxLength: 120
        description:
          type: string
          maxLength: 500
        technologies:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        url:
          type: string
          format: uri
      required:
      - id
...
    ResumeContent:
      type: object
      properties:
        personalInfo:
          $ref: '#/components/schemas/PersonalInfo'
        summary:
          type: string
          maxLength: 1000
        workExperience:
          type: array
          items:
            $ref: '#/components/schemas/WorkExperience'
        education:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        skills:
          type: array
          items:
            type: string
            maxLength: 60
          maxItems: 60
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
      required:
      - education
      - personalInfo
      - skills
      - workExperience
...
    TailorJDRequest:
      type: object
      properties:
        content:
          $ref: '#/components/schemas/ResumeContent'
        job_description:
          type: string
          maxLength: 8000
      required:
      - content
      - job_description
    TailorJDResponse:
      type: object
      properties:
        match_score:
          type: integer
        missing_keywords:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/TailorSuggestion'
        ai_usage:
          $ref: '#/components/schemas/AIUsage'
      required:
      - ai_usage
      - match_score
      - missing_keywords
      - suggestions
    TailorSuggestion:
      type: object
      properties:
        bullet_id:
          type: string
        current:
          type: string
        suggested:
          type: string
        adds:
          type: array
          items:
            type: string
      required:
      - adds
      - bullet_id
      - current
      - suggested
...
    WorkExperience:
      type: object
      properties:
        id:
          type: string
          maxLength: 64
        company:
          type: string
          maxLength: 120
        position:
          type: string
          maxLength: 120
        location:
          type: string
          maxLength: 120
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        bullets:
          type: array
          items:
            type: string
            maxLength: 500
          maxItems: 12
      required:
      - id
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT